Refactor(reconcile): Decompose engine
Split the reconciliation engine into focused package modules while
keeping the root import path and public compatibility surface stable.

The extracted planners preserve no-duplicate assignment safety and
existing diagnostics parity covered by the test suite.

Closes #627
